backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
djangorestframework_simplejwt==4.7.2
djoser==2.1.0
Pillow==9.3.0

# requests==2.26.0
# PyJWT==2.1.0
# pytest==6.2.4
# pytest-django==4.4.0
# pytest-pythonpath==0.7.3
# django-filter==23.5